Refactor PerformanceTrait

+++ src/performance/functional/quantize_linear/quantize_linear_fp/fp_i8/core.cairo
@@ -1,13 +1,14 @@
 use orion::numbers::fixed_point::core::{FixedType, FixedImpl};
 use orion::operators::tensor::core::{Tensor, ExtraParams};
-use orion::performance::functional::quantize_linear::quantize_linear_fp::fp_i32::fp8x23;
-use orion::performance::functional::quantize_linear::quantize_linear_fp::fp_i32::fp16x16;
+use orion::numbers::signed_integer::i8::i8;
+use orion::performance::functional::quantize_linear::quantize_linear_fp::fp_i8::fp8x23;
+use orion::performance::functional::quantize_linear::quantize_linear_fp::fp_i8::fp16x16;
 use orion::numbers::signed_integer::i32::i32;
 
 /// Cf: PerfomanceTrait::quantize_linear docstring
 fn quantize_linear(
     x: @Tensor<FixedType>, y_scale: @Tensor<FixedType>, y_zero_point: @Tensor<FixedType>
-) -> Option<Tensor<i32>> {
+) -> Option<Tensor<i8>> {
     match *x.extra {
         Option::Some(extra_params) => match extra_params.fixed_point {
             Option::Some(fixed_point) => match fixed_point {
